[android] Disable the app on the lock screen by default
Organic Maps drains the battery like crazy with default settings.

**Steps to reproduce:**

1. Install and run the app.
2. Lock your device and put it on a flat surface, with the screen facing up.
3. Pick up the device OR just tap on the screen.
4. The screen will automatically turn on, this is an Android feature.
5. Organic Maps will never allow the screen to go off after that.

Now, imaging this case happens while the phone is in a pocket. The battery may
The battery will completely drain before the user even notice that screen is on.
I see Organic Maps draining battery and burning screens on my devices regulary.

The problem is caused by two options which are enabled by default -
"Show on the lock screen" and "Allow the screen to sleep". The first option
adds the app to the lock screen (which is weird on its own). The second
option doesn't allow the phone's screen to go off. That is it.

This patch changes the default value of "Show on the lock screen" to OFF
to prevent this situation. It is still can be enabled manually by the user.

Signed-off-by: Roman Tsisyk <roman@tsisyk.com>